protoxform: Shift protoxform test to bazel and rename -> `protoprint`
